pyproject.toml
[project]
name = "fringe_app_v2"
version = "2.0.0"
description = "Structured-light 3D scanner — fringe projection, calibration, reconstruction, defect detection"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Henry"}]

dependencies = [
  "numpy>=1.24",
  "pygame>=2.5",
  "pyyaml>=6.0",
  "pillow>=10.0",
  "flask>=3.0",
  "scipy>=1.11",
  "opencv>=4.8",
]

[project.scripts]
fringe-app = "fringe_app_v2.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["fringe_app_v2*"]

[tool.pyright]
venvPath = "."
venv = ".venv"
